pythoncourse/requirements/development.txt
-r base.txt

gevent
gunicorn
django-s3-folder-storage==0.3
